+++ NEWS.md
@@ -5,6 +5,9 @@
 * Silenced an unknown attribute warning specific to the Intel compiler
   (r-lib/systemfonts#98).
 
+* `cpp_register()` gains an argument `extension=` governing the file extension of
+  the `src/cpp11` file. By default it's `.cpp`, but `.cc` is now supported as well (#292, @MichaelChirico)
+
 # cpp11 0.4.3
 
 * Modernized the GitHub Actions workflows and updated some internal tests to